# src/main/resources/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  version: 1.0.0
  description: API для управления банковскими картами, включая регистрацию, вход, обновление токенов, создание, просмотр, обновление и удаление карт, а также переводы между картами.
servers:
  - url: http://localhost:8080
    description: Локальный сервер
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          example: user1
        password:
          type: string
          format: password
          example: pass1234
        role:
          type: string
          enum:
            - USER
            - ADMIN
          example: USER
      required:
        - username
        - password
        - role
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: user1
        password:
          type: string
          format: password
          example: pass1234
      required:
        - username
        - password
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - accessToken
        - refreshToken
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        cardNumber:
          type: string
          description: Маскированный номер карты (например, **** **** **** 1234)
          example: "**** **** **** 1234"
        owner:
          type: string
          example: John Doe
        expiryDate:
          type: string
          format: date
          example: 2026-12-31
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balance:
          type: number
          format: double
          example: 1000.00
      required:
        - cardNumber
        - owner
        - expiryDate
        - status
        - balance
    CardDTO:
      type: object
      properties:
        cardNumber:
          type: string
          description: Полный номер карты для создания или обновления (будет зашифрован)
          example: "1234567812345678"
        owner:
          type: string
          example: John Doe
        expiryDate:
          type: string
          format: date
          example: 2026-12-31
        balance:
          type: number
          format: double
          example: 1000.00
      required:
        - cardNumber
        - owner
        - expiryDate
        - balance
    TransferDTO:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
          example: 1
        toCardId:
          type: integer
          format: int64
          example: 2
        amount:
          type: number
          format: double
          example: 500.00
      required:
        - fromCardId
        - toCardId
        - amount
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Error message
        status:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: 2025-08-08T15:16:00Z
paths:
  /user/register:
    post:
      summary: Регистрация пользователя
      operationId: registerUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Успешная регистрация с выдачей токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '500':
          description: Ошибка при регистрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/login:
    post:
      summary: Вход пользователя
      operationId: loginUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход с выдачей токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/refresh:
    post:
      summary: Обновление JWT токенов
      operationId: refreshToken
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Успешное обновление токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверный refresh токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /card/create:
    post:
      summary: Создание новой карты
      operationId: createCard
      tags:
        - Cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardDTO'
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: Неверные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещён (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /card/read:
    get:
      summary: Получение списка карт с пагинацией и фильтрацией
      operationId: readCards
      tags:
        - Cards
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
        - name: owner
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
                  pageable:
                    type: object
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /card/update:
    patch:
      summary: Обновление карты
      operationId: updateCard
      tags:
        - Cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardDTO'
      responses:
        '200':
          description: Карта успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: Неверные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещён (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /card/delete:
    delete:
      summary: Удаление карты
      operationId: deleteCard
      tags:
        - Cards
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта успешно удалена
        '400':
          description: Неверный идентификатор карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещён (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /card/transfer:
    post:
      summary: Перевод между картами
      operationId: transferMoney
      tags:
        - Cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferDTO'
      responses:
        '200':
          description: Перевод успешно выполнен
        '400':
          description: Неверные данные для перевода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещён (требуется роль USER)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /card/{id}/block:
    post:
      summary: Запрос на блокировку карты
      operationId: requestBlockCard
      tags:
        - Cards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Запрос на блокировку успешно отправлен
        '400':
          description: Неверный идентификатор карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещён (требуется роль USER)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
security:
  - BearerAuth: []